fix: split BigInt from ECC
