+++ ECC.cpp
@@ -7,201 +7,13 @@
 #include <gmp.h>
 
 #include "ECC.h"
+#include "BigInt.hpp"
 
 using std::string, std::vector, std::cout, std::endl;
 
 // ------------------- Type Definition -------------------
-class BigInt {
-private:
-    mpz_t value;
 
-public:
-    BigInt() {
-        mpz_init2(value, 256);
-    }
-
-    // hex here!
-    BigInt(const char* str) {
-        mpz_init2(value, 256);
-        mpz_set_str(value, str, 16);
-    }
-
-    BigInt(const int i) {
-        mpz_init2(value, 256);
-        mpz_set_si(value, i);
-    }
-
-    BigInt(const mpz_t& other) {
-        mpz_init2(value, 256);
-        mpz_set(value, other);
-    }
-
-    BigInt(const BigInt& other) {
-        mpz_init2(value, 256);
-        mpz_set(value, other.value);
-    }
-
-    ~BigInt() {
-        mpz_clear(value);
-    }
-
-    // ------------------- Operator Overloading -------------------
-    BigInt& operator=(const BigInt& other) {
-        if (this != &other) {
-            mpz_set(value, other.value);
-        }
-        return *this;
-    }
-
-    BigInt& operator=(const int i) {
-        mpz_set_si(value, i);
-        return *this;
-    }
-
-    BigInt operator+(const BigInt& other) const {
-        BigInt result;
-        mpz_add(result.value, value, other.value);
-        return result;
-    }
-
-    BigInt operator-(const BigInt& other) const {
-        BigInt result;
-        mpz_sub(result.value, value, other.value);
-        return result;
-    }
-
-    BigInt operator-() const {
-        BigInt result;
-        mpz_neg(result.value, value);
-        return result;
-    }
-
-    BigInt operator*(const BigInt& other) const {
-        BigInt result;
-        mpz_mul(result.value, value, other.value);
-        return result;
-    }
-
-    BigInt operator*(const int i) const {
-        BigInt result;
-        mpz_mul_si(result.value, value, i);
-        return result;
-    }
-
-    BigInt operator/(const BigInt& other) const {
-        BigInt result;
-        mpz_div(result.value, value, other.value);
-        return result;
-    }
-
-    BigInt operator%(const BigInt& other) const {
-        BigInt result;
-        mpz_mod(result.value, value, other.value);
-        return result;
-    }
-
-    BigInt operator^(const BigInt& other) const {
-        BigInt result;
-        mpz_xor(result.value, value, other.value);
-        return result;
-    }
-
-    BigInt operator+=(const BigInt& other) {
-        mpz_add(value, value, other.value);
-        return *this;
-    }
-
-    BigInt operator/= (const BigInt& other) {
-        mpz_div(value, value, other.value);
-        return *this;
-    }
-
-    BigInt operator%= (const BigInt& other) {
-        mpz_mod(value, value, other.value);
-        return *this;
-    }
-
-    BigInt operator++() {
-        mpz_add_ui(value, value, 1);
-        return *this;
-    }
-
-    bool operator==(const BigInt& other) const {
-        return mpz_cmp(value, other.value) == 0;
-    }
-
-    bool operator!=(const BigInt& other) const {
-        return mpz_cmp(value, other.value) != 0;
-    }
-
-    bool operator<(const BigInt& other) const {
-        return mpz_cmp(value, other.value) < 0;
-    }
-
-    bool operator<=(const BigInt& other) const {
-        return mpz_cmp(value, other.value) <= 0;
-    }
-
-    bool operator>(const BigInt& other) const {
-        return mpz_cmp(value, other.value) > 0;
-    }
-
-    bool operator>=(const BigInt& other) const {
-        return mpz_cmp(value, other.value) >= 0;
-    }
-
-
-
-    // ------------------- Member Function -------------------
-    mpz_t& getValue() {
-        return value;
-    }
-
-    void fromMessage(const string& message) {
-        mpz_import(this->value, message.size(), 1, sizeof(char), 0, 0, message.c_str());
-    }
-
-    string toMessage() const {
-        size_t count;
-        char* str = (char*)mpz_export(NULL, &count, 1, sizeof(char), 0, 0, value);
-        string result(str, count);
-        free(str);
-        return result;
-    }
-
-    string get16String() const {
-        char* str256bit = new char[65];
-        mpz_get_str(str256bit, 16, value);
-        string result(str256bit);
-        delete[] str256bit;
-        return result;
-    }
-
-    string get2String() const {
-        char* str256bit = new char[257];
-        mpz_get_str(str256bit, 2, value);
-        string result(str256bit);
-        delete[] str256bit;
-        return result;
-    }
-
-    static BigInt Random() {
-        mpz_t number;
-        mpz_init(number);
-
-        // Set the seed for random number generation
-        gmp_randstate_t state;
-        gmp_randinit_default(state);
-        gmp_randseed_ui(state, time(NULL));  // Set the seed value as desired
-
-        // Generate a random 256-bit number
-        mpz_urandomb(number, state, 256);
-
-        return BigInt(number);
-    }
-};
-
-typedef BigInt T;
+typedef BigInt<256> T;
 typedef struct Point {
     T x, y;
     Point() {}
@@ -216,7 +28,7 @@ typedef vector<Point> Messages;
 // G: compressed form
 static const T a = 0, b = 7;
 static const T field("FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEFFFFFC2F");
-static const Point G(BigInt("79BE667EF9DCBBAC55A06295CE870B07"), BigInt("029BFCDB2DCE28D959F2815B16F81798"));
+static const Point G(T("79BE667EF9DCBBAC55A06295CE870B07"), T("029BFCDB2DCE28D959F2815B16F81798"));
 
 // Block size, because secp256k1 is for 256-bit, so BLOCK_SIZE = 256 / 8 = 32
 const size_t BLOCK_SIZE = 32;
@@ -225,8 +37,6 @@ const size_t BLOCK_SIZE = 32;
 T getRandom() {
     return T::Random() % field;
 }
-
-
 
 // as + bt = gcd(a, b)
 void extendedEuclid(T a, T b, T& s, T& t, T& gcd) {
